change timestamp type from Int64 to double

+++ RelicModManager/DatabaseObject.cs
@@ -9,7 +9,7 @@ namespace RelhaxModpack
         //the developer's version of the mod
         public string version { get; set; }
         public string zipFile { get; set; }
-        public Int64 timestamp { get; set; }
+        public double timestamp { get; set; }
         public List<Config> configs = new List<Config>();
         //the start address of the zip file location. enabled us to use sites that
         //generate random filenames for ly shared files.
